21w42a

+++ src/main/kotlin/com/viaversion/aas/protocol/Util.kt
@@ -8,29 +8,29 @@ import com.viaversion.viaversion.api.type.Type
 
 val xyzToPosition = ValueReader { packetWrapper: PacketWrapper ->
     val x = packetWrapper.read(Type.INT)
-    val y = packetWrapper.read(Type.INT).toShort()
+    val y = packetWrapper.read(Type.INT)
     val z = packetWrapper.read(Type.INT)
     Position(x, y, z)
 }
 val xyzUBytePos = ValueReader { packetWrapper: PacketWrapper ->
     val x = packetWrapper.read(Type.INT)
-    val y = packetWrapper.read(Type.UNSIGNED_BYTE)
+    val y = packetWrapper.read(Type.UNSIGNED_BYTE).toInt()
     val z = packetWrapper.read(Type.INT)
     Position(x, y, z)
 }
 val xyzUBytePosWriter = ValueWriter { packetWrapper: PacketWrapper, pos: Position ->
-    packetWrapper.write(Type.INT, pos.x)
-    packetWrapper.write(Type.UNSIGNED_BYTE, pos.y.toShort())
-    packetWrapper.write(Type.INT, pos.z)
+    packetWrapper.write(Type.INT, pos.x())
+    packetWrapper.write(Type.UNSIGNED_BYTE, pos.y().toShort())
+    packetWrapper.write(Type.INT, pos.z())
 }
 val xyzShortPosWriter = ValueWriter { packetWrapper: PacketWrapper, pos: Position ->
-    packetWrapper.write(Type.INT, pos.x)
-    packetWrapper.write(Type.SHORT, pos.y.toShort())
-    packetWrapper.write(Type.INT, pos.z)
+    packetWrapper.write(Type.INT, pos.x())
+    packetWrapper.write(Type.SHORT, pos.y().toShort())
+    packetWrapper.write(Type.INT, pos.z())
 }
 val xyzShortPos = ValueReader { packetWrapper: PacketWrapper ->
     val x = packetWrapper.read(Type.INT)
-    val y = packetWrapper.read(Type.SHORT)
+    val y = packetWrapper.read(Type.SHORT).toInt()
     val z = packetWrapper.read(Type.INT)
     Position(x, y, z)
 }
